refactor: data/visit 패키지명을 moment로 변경
